test.py: Move get configured modes to common lib
This will allow using this method inside the test module for pytest launching the boost and unit tests
