+++ test/pylib/util.py
@@ -3,6 +3,8 @@
 #
 # SPDX-License-Identifier: LicenseRef-ScyllaDB-Source-Available-1.0
 #
+import re
+import subprocess
 from collections.abc import Coroutine
 import threading
 import time
@@ -10,7 +12,8 @@ import asyncio
 import logging
 import pathlib
 import os
-import pytest
+from functools import cache
+
 import random
 import string
 
@@ -255,3 +258,21 @@ async def wait_for_first_completed(coros: list[Coroutine]):
         t.cancel()
     for t in done:
         await t
+
+
+def ninja(target):
+    """Build specified target using ninja"""
+    build_dir = 'build'
+    args = ['ninja', target]
+    if os.path.exists(os.path.join(build_dir, 'build.ninja')):
+        args = ['ninja', '-C', build_dir, target]
+    return subprocess.Popen(args, stdout=subprocess.PIPE).communicate()[0].decode()
+
+
+@cache
+def get_configured_modes():
+    out = ninja('mode_list')
+    # [1/1] List configured modes
+    # debug release dev
+    return re.sub(r'.* List configured modes\n(.*)\n', r'\1',
+                            out, count=1, flags=re.DOTALL).split('\n')[-1].split(' ')
